WIP net messages

+++ src/SourceDemoParser.Net/Frames/Classes/Net/Messages/NetTickMessage.cs
@@ -3,7 +3,7 @@ namespace SourceDemoParser
 	public class NetTickMessage : INetMessage
 	{
 		public int Tick { get; set; }
-		public uint HostFrameTime { get; set; }
-		public uint HostFrameTimeStdDeviation { get; set; }
+		public short HostFrameTime { get; set; }
+		public short HostFrameTimeStdDeviation { get; set; }
 	}
 }
