+++ src/SourceDemoParser.Net/Frames/Classes/Net/Messages/SvcUpdateStringTableMessage.cs
@@ -2,10 +2,10 @@ namespace SourceDemoParser
 {
 	public class SvcUpdateStringTableMessage : INetMessage
 	{
-		public uint Id { get; set; }
+		public int Id { get; set; }
 		public bool EntriesChanged { get; set; }
-		public ushort ChangedEntries { get; set; }
-		public uint Length { get; set; }
+		public int ChangedEntries { get; set; }
+		public int Length { get; set; }
 		public byte[] Data { get; set; }
 	}
 }
